modelForEventController fixed spring security varsion 1.0.jar is ready for deployment! Alhamdolillah V1.0. Bug fixes. Custom Login form and logout is readye exception handling for 403, 404, 500 resource init Soap notif service init soap working without unmarshalling app now is in a pretty Good shape. Deployment: V1.0 Otp init. im still confused UI is rather ready. Thank you, GOD OTP init. May GOD help us Otp phases: 1. Generate and verify OTP and its details (OK) phase2. Authentication (init) Phase2(OK-Thank you, GOD) - Remaining - some ui messages base template edited for deployment TaskServiceImpl is getting better Pm and Event revision init: Controller params, Model, Forms, templates... Re-Building Pm - i think the TaskDetail update will work fine PmController through service and repository updated

+++ src/main/java/ir/tic/clouddc/pm/PmInterfaceRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PmInterfaceRepository extends JpaRepository<PmInterface, Integer> {
+public interface PmInterfaceRepository extends JpaRepository<PmInterface, Short> {
     @Query("SELECT p.id FROM PmInterface p WHERE p.type.id = :typeId")
     List<Pm> fetchRelatedPmList(@Param("typeId") int typeId);
